+++ src/zepben/evolve/model/busbranch/bus_branch.py
@@ -901,7 +901,6 @@ async def _group_negligible_impedance_terminals(
     trace = BasicTraversal(
         start_item=terminal,
         queue_next=_queue_terminals_across_negligible_impedance(has_negligible_impedance),
-        process_queue=LifoQueue(),
         step_actions=[_process_terminal(tg, has_negligible_impedance)]
     )
     await trace.run()
@@ -946,7 +945,6 @@ async def _group_common_ac_line_segment_terminals(acls: AcLineSegment) -> Termin
     trace = BasicTraversal(
         start_item=acls,
         queue_next=_queue_common_impedance_lines(common_acls, has_common_impedance),
-        process_queue=LifoQueue(),
         step_actions=[_process_acls(common_acls, connectivity_node_counter)]
     )
     await trace.run()